+++ algosdk/encoding.py
@@ -26,9 +26,32 @@ def msgpack_encode(obj):
         the most recent version of msgpack rather than the older msgpack
         version that had no "bin" family).
     """
-    if not isinstance(obj, OrderedDict):
-        obj = obj.dictify()
-    return base64.b64encode(msgpack.packb(obj, use_bin_type=True)).decode()
+    d = obj
+    if not isinstance(obj, dict):
+        d = obj.dictify()
+    od = _sort_dict(d)
+    return base64.b64encode(msgpack.packb(od, use_bin_type=True)).decode()
+
+
+def _sort_dict(d):
+    """
+    Sorts a dictionary recursively and removes all zero values.
+
+    Args:
+        d (dict): dictionary to be sorted
+
+    Returns:
+        OrderedDict: sorted dictionary with no zero values
+    """
+    od = OrderedDict()
+    for k, v in sorted(d.items()):
+        if isinstance(v, dict):
+            od[k] = _sort_dict(v)
+        else:
+            od[k] = v
+        if not od[k]:
+            del od[k]
+    return od
 
 
 def msgpack_decode(enc):